test(properties): Improved testing of PathProperties. Help doc for PathProperties is now also consistent.
