+++ configuration/src/main/java/com/synopsys/integration/configuration/property/types/path/PathValue.kt
@@ -1,0 +1,13 @@
+package com.synopsys.integration.configuration.property.types.path
+
+import java.nio.file.Path
+
+data class PathValue(private val value: String) {
+    fun resolvePath(pathResolver: PathResolver): Path {
+        return pathResolver.resolvePath(value)
+    }
+
+    override fun toString(): String {
+        return value
+    }
+}
